CommitFinalBlogPessoal

+++ Spring/blog-Pessoal/src/main/java/org/generation/blog/Pessoal/controller/TemaController.java
@@ -2,6 +2,9 @@
 
 import java.util.List;
 
+import javax.validation.Valid;
+
+
 import org.generation.blog.Pessoal.model.Tema;
 import org.generation.blog.Pessoal.repository.TemaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,7 +18,11 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
+
+import net.bytebuddy.dynamic.DynamicType.Builder.FieldDefinition.Optional;
 
 
 @RestController
@@ -39,16 +46,44 @@ public class TemaController {
 		return ResponseEntity.ok(repository.findAllByDescricaoContainingIgnoreCase(nome));
 	}
 	@PostMapping
-	public ResponseEntity<Tema> post 	(@RequestBody Tema tema){
-		return ResponseEntity.status(HttpStatus.CREATED)
-				.body(repository.save(tema));
+
+	public ResponseEntity<Tema> post(@Valid @RequestBody Tema tema){
+
+	        return ResponseEntity.status(HttpStatus.CREATED)
+
+			        .body(repository.save(tema));
+
 	}
+
+
+
 	@PutMapping
-	public ResponseEntity<Tema> put (@RequestBody Tema tema){
-		return ResponseEntity.ok(repository.save(tema));
+
+	public ResponseEntity<Tema> put(@Valid @RequestBody Tema tema){
+
+	        return repository.findById(tema.getId())
+
+			        .map(resp -> ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema)))
+
+			        .orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
+
 	}
+
+
+
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+
 	@DeleteMapping("/{id}")
+
 	public void delete(@PathVariable long id) {
-		repository.deleteById(id);
+
+	        java.util.Optional<Tema> tema = repository.findById(id);
+
+	        if(tema.isEmpty())
+
+		        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
+	        repository.deleteById(id);				
+
 	}
 }
